HBASE-27830 Introduce hdfs overlay
Provide a basic implementation for running HDFS using the docker image provided in
HBASE-27846. Also provides testing capabilities using the test image added in HBASE-27829. Makes
some tweaks to that image to get it running in a linux environment -- Docker on MacOS make some
user permission adjustments that are not present in linux.